Admin tool: change root package and create binary distributions
