+++ quarkus/admin/src/main/java/org/apache/polaris/admintool/BaseCommand.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.quarkus.admin;
+package org.apache.polaris.admintool;
 
 import jakarta.inject.Inject;
 import java.util.concurrent.Callable;
