+++ quarkus/admin/src/main/java/org/apache/polaris/admintool/BootstrapCommand.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.quarkus.admin;
+package org.apache.polaris.admintool;
 
 import java.util.List;
 import java.util.Map;
